Revert "Move common forwarding rules comparison function to utils."
This reverts commit 6dcfa1a577a0ee716fb78cc6fdece5540ed4445e.
