+++ pkg/utils/utils_test.go
@@ -20,14 +20,12 @@ import (
 	"context"
 	"errors"
 	"fmt"
+	"k8s.io/klog/v2"
 	"net/http"
 	"reflect"
 	"testing"
 	"time"
 
-	"k8s.io/klog/v2"
-
-	"github.com/GoogleCloudPlatform/k8s-cloud-provider/pkg/cloud"
 	"github.com/google/go-cmp/cmp"
 	"google.golang.org/api/compute/v1"
 	"google.golang.org/api/googleapi"
@@ -41,7 +39,6 @@ import (
 	"k8s.io/apimachinery/pkg/types"
 	"k8s.io/cloud-provider-gcp/providers/gce"
 	"k8s.io/ingress-gce/pkg/annotations"
-	"k8s.io/ingress-gce/pkg/composite"
 	"k8s.io/ingress-gce/pkg/flags"
 	"k8s.io/ingress-gce/pkg/utils/common"
 )
@@ -242,243 +239,6 @@ func TestEqualResourceIDs(t *testing.T) {
 		t.Run(name, func(t *testing.T) {
 			if got := EqualResourceIDs(tc.a, tc.b); got != tc.want {
 				t.Errorf("EqualResourceIDs(%q, %q) = %v, want %v", tc.a, tc.b, got, tc.want)
-			}
-		})
-	}
-}
-
-func TestEqualForwardingRules(t *testing.T) {
-	t.Parallel()
-
-	fwdRules := []*composite.ForwardingRule{
-		{
-			Name:                "empty-ip-address-fwd-rule",
-			IPAddress:           "",
-			Ports:               []string{"123"},
-			IPProtocol:          "TCP",
-			LoadBalancingScheme: string(cloud.SchemeInternal),
-			BackendService:      "http://www.googleapis.com/projects/test/regions/us-central1/backendServices/bs1",
-		},
-		{
-			Name:                "tcp-fwd-rule",
-			IPAddress:           "10.0.0.0",
-			Ports:               []string{"123"},
-			IPProtocol:          "TCP",
-			LoadBalancingScheme: string(cloud.SchemeInternal),
-			BackendService:      "http://www.googleapis.com/projects/test/regions/us-central1/backendServices/bs1",
-		},
-		{
-			Name:                "udp-fwd-rule",
-			IPAddress:           "10.0.0.0",
-			Ports:               []string{"123"},
-			IPProtocol:          "UDP",
-			LoadBalancingScheme: string(cloud.SchemeInternal),
-			BackendService:      "http://www.googleapis.com/projects/test/regions/us-central1/backendServices/bs1",
-		},
-		{
-			Name:                "global-access-fwd-rule",
-			IPAddress:           "10.0.0.0",
-			Ports:               []string{"123"},
-			IPProtocol:          "TCP",
-			LoadBalancingScheme: string(cloud.SchemeInternal),
-			AllowGlobalAccess:   true,
-			BackendService:      "http://www.googleapis.com/projects/test/regions/us-central1/backendServices/bs1",
-		},
-		{
-			Name:                "fwd-rule-bs-link1",
-			IPAddress:           "10.0.0.0",
-			Ports:               []string{"123"},
-			IPProtocol:          "TCP",
-			LoadBalancingScheme: string(cloud.SchemeInternal),
-			BackendService:      "http://compute.googleapis.com/projects/test/regions/us-central1/backendServices/bs1",
-		},
-		{
-			Name:                "fwd-rule-bs-link2",
-			IPAddress:           "10.0.0.0",
-			Ports:               []string{"123"},
-			IPProtocol:          "TCP",
-			LoadBalancingScheme: string(cloud.SchemeInternal),
-			BackendService:      "http://www.googleapis.com/projects/test/regions/us-central1/backendServices/bs1",
-		},
-		{
-			Name:                "udp-fwd-rule-all-ports",
-			IPAddress:           "10.0.0.0",
-			Ports:               []string{"123"},
-			AllPorts:            true,
-			IPProtocol:          "UDP",
-			LoadBalancingScheme: string(cloud.SchemeInternal),
-			BackendService:      "http://www.googleapis.com/projects/test/regions/us-central1/backendServices/bs1",
-			NetworkTier:         cloud.NetworkTierPremium.ToGCEValue(),
-		},
-		{
-			Name:                "fwd-rule-bs-link2-standard-ntier",
-			IPAddress:           "10.0.0.0",
-			Ports:               []string{"123"},
-			IPProtocol:          "TCP",
-			LoadBalancingScheme: string(cloud.SchemeInternal),
-			BackendService:      "http://www.googleapis.com/projects/test/regions/us-central1/backendServices/bs1",
-			NetworkTier:         string(cloud.NetworkTierStandard),
-		},
-		{
-			Name:                "fwd-rule-bs-link2-premium-ntier",
-			IPAddress:           "10.0.0.0",
-			Ports:               []string{"123"},
-			IPProtocol:          "TCP",
-			LoadBalancingScheme: string(cloud.SchemeInternal),
-			BackendService:      "http://www.googleapis.com/projects/test/regions/us-central1/backendServices/bs1",
-			NetworkTier:         cloud.NetworkTierPremium.ToGCEValue(),
-		},
-	}
-
-	frPortRange1 := &composite.ForwardingRule{
-		Name:                "tcp-fwd-rule",
-		IPAddress:           "10.0.0.0",
-		PortRange:           "2-3",
-		IPProtocol:          "TCP",
-		LoadBalancingScheme: string(cloud.SchemeInternal),
-		BackendService:      "http://www.googleapis.com/projects/test/regions/us-central1/backendServices/bs1",
-	}
-	frPortRange2 := &composite.ForwardingRule{
-		Name:                "tcp-fwd-rule",
-		IPAddress:           "10.0.0.0",
-		PortRange:           "1-2",
-		IPProtocol:          "TCP",
-		LoadBalancingScheme: string(cloud.SchemeInternal),
-		BackendService:      "http://www.googleapis.com/projects/test/regions/us-central1/backendServices/bs1",
-	}
-
-	for _, tc := range []struct {
-		desc        string
-		oldFwdRule  *composite.ForwardingRule
-		newFwdRule  *composite.ForwardingRule
-		expectEqual bool
-	}{
-		{
-			desc:        "empty ip address does not match valid ip",
-			oldFwdRule:  fwdRules[0],
-			newFwdRule:  fwdRules[1],
-			expectEqual: false,
-		},
-		{
-			desc:        "global access enabled",
-			oldFwdRule:  fwdRules[1],
-			newFwdRule:  fwdRules[3],
-			expectEqual: false,
-		},
-		{
-			desc:        "IP protocol changed",
-			oldFwdRule:  fwdRules[1],
-			newFwdRule:  fwdRules[2],
-			expectEqual: false,
-		},
-		{
-			desc:        "same forwarding rule",
-			oldFwdRule:  fwdRules[3],
-			newFwdRule:  fwdRules[3],
-			expectEqual: true,
-		},
-		{
-			desc:        "same forwarding rule, different basepath",
-			oldFwdRule:  fwdRules[4],
-			newFwdRule:  fwdRules[5],
-			expectEqual: true,
-		},
-		{
-			desc:        "same forwarding rule, one uses ALL keyword for ports",
-			oldFwdRule:  fwdRules[2],
-			newFwdRule:  fwdRules[6],
-			expectEqual: false,
-		},
-		{
-			desc:        "network tier mismatch",
-			oldFwdRule:  fwdRules[6],
-			newFwdRule:  fwdRules[7],
-			expectEqual: false,
-		},
-		{
-			desc:        "same forwarding rule, different port ranges",
-			oldFwdRule:  frPortRange1,
-			newFwdRule:  frPortRange2,
-			expectEqual: false,
-		},
-		{
-			desc: "network mismatch",
-			oldFwdRule: &composite.ForwardingRule{
-				Name:                "tcp-fwd-rule",
-				IPAddress:           "10.0.0.0",
-				Ports:               []string{"123"},
-				IPProtocol:          "TCP",
-				LoadBalancingScheme: string(cloud.SchemeInternal),
-				BackendService:      "http://www.googleapis.com/projects/test/regions/us-central1/backendServices/bs1",
-				Network:             "https://www.googleapis.com/compute/v1/projects/test-poject/global/networks/test-vpc",
-			},
-			newFwdRule: &composite.ForwardingRule{
-				Name:                "tcp-fwd-rule",
-				IPAddress:           "10.0.0.0",
-				Ports:               []string{"123"},
-				IPProtocol:          "TCP",
-				LoadBalancingScheme: string(cloud.SchemeInternal),
-				BackendService:      "http://www.googleapis.com/projects/test/regions/us-central1/backendServices/bs1",
-				Network:             "https://www.googleapis.com/compute/v1/projects/test-poject/global/networks/test-other-vpc",
-			},
-			expectEqual: false,
-		},
-		{
-			desc: "subnetwork mismatch",
-			oldFwdRule: &composite.ForwardingRule{
-				Name:                "tcp-fwd-rule",
-				IPAddress:           "10.0.0.0",
-				Ports:               []string{"123"},
-				IPProtocol:          "TCP",
-				LoadBalancingScheme: string(cloud.SchemeInternal),
-				BackendService:      "http://www.googleapis.com/projects/test/regions/us-central1/backendServices/bs1",
-				Network:             "https://www.googleapis.com/compute/v1/projects/test-poject/global/networks/test-vpc",
-				Subnetwork:          "https://www.googleapis.com/compute/v1/projects/test-poject/regions/us-central1/subnetworks/default-subnet",
-			},
-			newFwdRule: &composite.ForwardingRule{
-				Name:                "tcp-fwd-rule",
-				IPAddress:           "10.0.0.0",
-				Ports:               []string{"123"},
-				IPProtocol:          "TCP",
-				LoadBalancingScheme: string(cloud.SchemeInternal),
-				BackendService:      "http://www.googleapis.com/projects/test/regions/us-central1/backendServices/bs1",
-				Network:             "https://www.googleapis.com/compute/v1/projects/test-poject/global/networks/test-vpc",
-				Subnetwork:          "https://www.googleapis.com/compute/v1/projects/test-poject/regions/us-central1/subnetworks/other-subnet",
-			},
-			expectEqual: false,
-		},
-		{
-			desc: "equal network data",
-			oldFwdRule: &composite.ForwardingRule{
-				Name:                "tcp-fwd-rule",
-				IPAddress:           "10.0.0.0",
-				Ports:               []string{"123"},
-				IPProtocol:          "TCP",
-				LoadBalancingScheme: string(cloud.SchemeInternal),
-				BackendService:      "http://www.googleapis.com/projects/test/regions/us-central1/backendServices/bs1",
-				Network:             "https://www.googleapis.com/compute/v1/projects/test-poject/global/networks/test-vpc",
-				Subnetwork:          "https://www.googleapis.com/compute/v1/projects/test-poject/regions/us-central1/subnetworks/default-subnet",
-			},
-			newFwdRule: &composite.ForwardingRule{
-				Name:                "tcp-fwd-rule",
-				IPAddress:           "10.0.0.0",
-				Ports:               []string{"123"},
-				IPProtocol:          "TCP",
-				LoadBalancingScheme: string(cloud.SchemeInternal),
-				BackendService:      "http://www.googleapis.com/projects/test/regions/us-central1/backendServices/bs1",
-				Network:             "projects/test-poject/global/networks/test-vpc",
-				Subnetwork:          "projects/test-poject/regions/us-central1/subnetworks/default-subnet",
-			},
-			expectEqual: true,
-		},
-	} {
-		t.Run(tc.desc, func(t *testing.T) {
-			got, err := EqualForwardingRules(tc.oldFwdRule, tc.newFwdRule)
-			if err != nil {
-				t.Errorf("EqualForwardingRules(_, _) = %v, want nil error", err)
-			}
-			if got != tc.expectEqual {
-				t.Errorf("EqualForwardingRules(_, _) = %t, want %t", got, tc.expectEqual)
 			}
 		})
 	}
